To-do list
This is an in progress project that revolves around user convenience. Currently the user can create, edit, and delete a To-Do List.  The user friendly aspects can be seen what the user loads or deletes a file without a file-chooser interface. This allows the user to quickly load lists, and delete whatever list they need to. Future updates will aim to make this application even more user friendly.